Apply these changes
TAB 1
Instead of "Chat with SciLifeLab websites" I would say "Ask a question about SciLifeLab"
Then under the title I would add a description, something like "This is a tool that can answer questions about SciLifeLab based on the information that is available on scilifelab.se website on top of the general knowledge of a Large Language Model (in this case ChatGPT 3.5). You can ask one question at a time but the question can be as detailed/as long as you wish."
When a user sends a new question I would remove the old one so that it is clear that the history is not saved.
The instead of "Submit" I would have the button text be "Ask".
I would remove the "Chatbot" text on this page if possible.
TAB 2
Instead of "General LLMs" I would say "Ask general LLMs"
There under I would also add a description text "Here you can ask a question to one of the general LLMs in the list below (currently two versions of ChatGPT). You can compare the answers you get from these two LLMs and from the specialized SciLifeLab tool in the first tab."
I would remove the "Chatbot" text on this page if possible.